% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centile_table.R
\name{calculate_centile_table}
\alias{calculate_centile_table}
\title{Calculate a centile table from a given reference}
\usage{
calculate_centile_table(x, z = NULL, p = NULL, refcode = NULL, dec = 4, ...)
}
\arguments{
\item{x}{A vector with the grid (usually years, sometimes cm) on the
horizontal axis.}

\item{z}{A vector containing the z values of the desired centiles.}

\item{p}{A vector containing the desired percentile values (between 0 and
100). The default \code{p = NULL} does not calculate percentiles.}

\item{refcode}{A reference code according to the specifications of the
\code{centile} package. This code should match a pre-defined reference table
containing LMS values for various age or measurement groups.}

\item{dec}{Number of decimals needed for the output.}

\item{\dots}{Additional arguments passed down to \code{z2y()}.}
}
\value{
A \code{matrix} containing the desired centiles, with columns named
according to either the z-score values (e.g., \code{Z-2}, \code{Z0}, \code{Z2}) or
percentiles (e.g., \code{P3}, \code{P50}, \code{P97}).
}
\description{
This function calculates centiles based on specified z-scores or percentiles
using a given reference. It requires the reference to be provided as a code
compatible with the \code{centile} package.
}
\examples{
# Centile table from WHO weight standard for boys aged 0-5 years
calculate_centile_table(z = -2:2, x = 0:5, refcode = "who_2006_wgt_male_")
}
