% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rif.R
\name{import_rif}
\alias{import_rif}
\title{Import a RIF formatted file into a tibble}
\usage{
import_rif(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. To be recognised as
literal data, the input must be either wrapped with \code{I()}, be a string
containing at least one new line, or be a vector containing at least one
string with a new line.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}
}
\value{
An object of class \code{reference}, basically a tibble with an attribute
called \code{study}.
}
\description{
Reads a file that is stored in RIF format and returns a tibble. The function
validates the presence of required fields, and checks whether the variable
names in the data match the specified distribution.
}
\examples{
fn <- system.file("testdata/nl_2009_hgt_male_nl.txt", package = "centile")
myref <- import_rif(fn)
}
\author{
Stef van Buuren 2021
}
