% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_reference.R
\name{is_reference}
\alias{is_reference}
\title{Test whether object is a reference}
\usage{
is_reference(x)
}
\arguments{
\item{x}{An R object, typically produced by \code{import_rif()}. The object is
a data frame with an attribute named \code{study}.}
}
\value{
Logical
}
\description{
Test whether object is a reference
}
\examples{
is_reference("this is a string")

# import reference and check
fn <- system.file("testdata/nl_2009_hgt_female_nl.txt", package = "centile")
myref <- import_rif(fn)
is_reference(myref)
}
