% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference.R
\name{load_reference}
\alias{load_reference}
\title{Load growth reference}
\usage{
load_reference(
  refcode = NULL,
  element = c("all", "table", "index", "study"),
  pkg = "centile",
  verbose = FALSE
)
}
\arguments{
\item{refcode}{String, code of a reference. Only the first element is loaded.}

\item{element}{Keyword, either \code{"all"}, \code{"table"}, \code{"index"} or \code{"study"}.
The default is \code{"all"}.}

\item{pkg}{The package containing the reference. The package must be loaded
and attached. The default \code{pkg = "centile"} searches in the home package.}

\item{verbose}{For more verbose output, set to \code{TRUE}.}
}
\value{
The return value depends on the \code{element} parameter.\tabular{ll}{
   \code{element} \tab Return value \cr
   \code{"all"} \tab All stored information \cr
   \code{"table"} \tab A \code{tibble} with reference values \cr
   \code{"index"} \tab Numeric vector with index values \cr
   \code{"study"} \tab Named character vector with study data \cr
}
}
\description{
This function provides access to the built-in growth references.
}
\examples{
ref <- load_reference("gc_2019_dsc_male_")
}
