% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_agegrid.R
\name{make_agegrid}
\alias{make_agegrid}
\title{Creates an age grid according to a specified format}
\usage{
make_agegrid(
  grid = c("compact", "classic", "extensive", "0-104w", "0-24m", "0-21y"),
  unit = c("year", "month", "week", "day"),
  dec = 4L
)
}
\arguments{
\item{grid}{A character string specifying one of the following:
\code{"compact"}, \code{"classic"}, \code{"extensive"}, \code{"0-104w"},
\code{"0-24m"} and \code{"0-21y"}.  The
default produces an age grid between 0 and 21
years with 95 points.}

\item{unit}{Character string specifying the time unit. Choices
are \code{"year"}, \code{"month"}, \code{"week"} and \code{"day"}.}

\item{dec}{Number of decimals for the result.}
}
\value{
Numeric vector or \code{NULL}
}
\description{
Creates an age grid according to a specified format
}
\examples{
head(make_agegrid("classic"))
}
\author{
Stef van Buuren, 2021
}
