% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2z.R
\name{p2z}
\alias{p2z}
\title{Convert centiles into Z-scores}
\usage{
p2z(p, scale = 100)
}
\arguments{
\item{p}{A numerical vector with centiles.}

\item{scale}{The maximum of the scale. The default (\code{100}) returns for
percentiles. Set \code{scale = 1} to obtain probabilities.}
}
\value{
A vector with \code{length(p)} elements containing centiles.
}
\description{
Functions \code{z2p()} and \code{p2z()} functions are the inverse of each other.
}
\examples{
p <- c(2.5, 10, 50, 90, 97.5)
p2z(p)
}
\author{
Stef van Buuren, 2021
}
