% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y.R
\name{y}
\alias{y}
\title{Convert Z-scores into measurements}
\usage{
y(
  z,
  x,
  refcode,
  pkg = "yzy",
  verbose = FALSE,
  dec = 3L,
  rule = 1L,
  tail_adjust = FALSE,
  ...
)
}
\arguments{
\item{z}{A numerical vector with Z-scores.}

\item{x}{A vector containing \code{length(y)} values of the numerical covariate (typically
decimal age or height) at which conversion is desired. A scalar \code{x} will be
expanded to \code{length(y)}.}

\item{refcode}{A character vector with \code{length(y)} elements, each of which
points to a reference. A scalar \code{refcode} will be expanded to \code{length(y)}.}

\item{pkg}{The package containing the references. The package must be loaded.
The default \code{pkg = "yzy"} searches in the home package.}

\item{verbose}{Set to \code{TRUE} to turn on warnings}

\item{dec}{A scalar value indicating the number of decimals used to round the
value. The default is 3 decimals.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation
    is to take place outside the interval [\code{min(x)}, \code{max(x)}].
    If \code{rule} is \code{1} then \code{NA}s are returned for such
    points and if it is \code{2}, the value at the closest data extreme
    is used.  Use, e.g., \code{rule = 2:1}, if the left and right side
    extrapolation should differ.}

\item{tail_adjust}{Logical. If \code{TRUE} the procedure applies the WHO
method for tail adjustment under the LMS distribution.
The default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
A vector with \code{length(z)} elements containing the measurements.
}
\description{
Converts age- or hgt-conditional standard deviation scores into measurements
(KG, CM, ...) using one or more external reference distributions.
}
\note{
The type of reference distribution is defined by the \code{distribution} field
of the \code{study} attribute in the reference. The function executes
transformations specified by the \code{tx} field (before calculation of \code{y}) and
the \code{yt} field (after calculating of \code{y}), so that the \code{y} have the original
units even if the reference is based on some transform of \code{y}.

Functions \code{z()} and \code{y()} functions are the inverse of each other.
}
\examples{
z <- c(0, -1.662, 0.928, -1.662, -1.268, -1.697, -1.268, -1.697)
x <- rep(c(0.3, 60), 4)
refcode <- c(
  "nl_2012_hdc_male_20", "nl_2012_wfh_male_",
  "nl_2012_hdc_male_30", "nl_2012_wfh_male_",
  "nl_1997_hdc_male_nl", "nl_1997_wfh_male_nla",
  "nl_1997_hdc_male_nl", "nl_1997_wfh_male_nla"
)
y <- y(z, x, refcode)
}
\author{
Stef van Buuren, 2021
}
