% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z}
\alias{z}
\title{Convert measurements into Z-scores}
\usage{
z(y, x, refcode, pkg = "yzy", dec = 3L, tail_adjust = FALSE, ...)
}
\arguments{
\item{y}{A numerical vector with measurements`.}

\item{x}{A vector containing \code{length(y)} values of the numerical covariate (typically
decimal age or height) at which conversion is desired.}

\item{refcode}{A character vector with \code{length(y)} elements, each of which
points to a reference.}

\item{pkg}{The package containing the reference. The package must be loaded
and attached. The default \code{pkg = "yzy"} searches in the home package.}

\item{dec}{A scalar value indicating the number of decimals used to round the
value. The default is 3 decimals.}

\item{tail_adjust}{Logical. If \code{TRUE} the procedure applies the WHO
method for tail adjustment under the LMS distribution.
The default is \code{FALSE}.}

\item{\dots}{Not used.}
}
\value{
A vector with \code{length(y)} elements containing the Z-scores.
}
\description{
Converts measurements into age- or hgt-conditional standard deviation scores
(SDS, Z-score) using one or more external reference distributions.
}
\note{
The type of reference distribution is defined by the \code{distribution} field
of the \code{study} attribute in the reference. The function executes
transformations specified by the \code{tx} and \code{ty} fields before calculating the
Z-score.

Functions \code{z()} and \code{y()} functions are the inverse of each other.
}
\examples{
y <- c(50, 50, 60, 60, 4, 4, 4, 4)
x <- c(rep(0.1, 4), rep(0.1, 4))
refcode <- c(rep(c("who_2011_hgt_male_", "who_2011_hgt_female_"), 2),
             rep(c("who_2011_wgt_male_", "who_2011_wgt_female_"), 2))
z(y, x, refcode)
}
\author{
Stef van Buuren, 2021
}
