% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2p.R
\name{z2p}
\alias{z2p}
\title{Convert Z-scores into centiles}
\usage{
z2p(z, scale = 100)
}
\arguments{
\item{z}{A numerical vector with Z-scores.}

\item{scale}{The maximum of the scale. The default (\code{100}) returns for
percentiles. Set \code{scale = 1} to obtain probabilities.}
}
\value{
A vector with \code{length(z)} elements containing centiles.
}
\description{
Functions \code{z2p()} and \code{p2z()} functions are the inverse of each other.
}
\examples{
z <- c(-1.96, -1.28, 0, 1.28, 1.96)
z2p(z)
}
\author{
Stef van Buuren, 2021
}
