% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2y.R
\name{z2y}
\alias{z2y}
\title{Convert Z-scores into measurements}
\usage{
z2y(
  z,
  x,
  refcode,
  pkg = "centile",
  verbose = FALSE,
  dec = 3L,
  rule = 1L,
  tail_adjust = FALSE,
  ...
)
}
\arguments{
\item{z}{A numerical vector with Z-scores.}

\item{x}{A vector containing \code{length(y)} values of the numerical covariate (typically
decimal age or height) at which conversion is desired. A scalar \code{x} will be
expanded to \code{length(y)}.}

\item{refcode}{A character vector with \code{length(y)} elements, each of which
is the name of a reference. Scalar \code{refcode} expands to \code{length(y)}.
Alternatively, \code{refcode} may be an object of class \code{reference}, typically
produced by \code{\link[=import_rif]{import_rif()}}.}

\item{pkg}{The package containing the references in the \code{R/sysdata.dta} object.
The package needs to be loaded. The default \code{pkg = "centile"} searches in
the \code{centile} package.}

\item{verbose}{Set to \code{TRUE} to turn on warnings}

\item{dec}{A scalar value indicating the number of decimals used to round the
value. The default is 3 decimals.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation
    is to take place outside the interval [\code{min(x)}, \code{max(x)}].
    If \code{rule} is \code{1} then \code{NA}s are returned for such
    points and if it is \code{2}, the value at the closest data extreme
    is used.  Use, e.g., \code{rule = 2:1}, if the left and right side
    extrapolation should differ.}

\item{tail_adjust}{Logical. If \code{TRUE} the procedure applies the WHO
method for tail adjustment under the LMS distribution. We do not recommend
this. The default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
A vector with \code{length(z)} elements containing the measurements.
}
\description{
Converts age- or hgt-conditional standard deviation scores into measurements
(KG, CM, ...) using one or more external reference distributions.
}
\note{
The type of reference distribution is defined by the \code{distribution} field
of the \code{study} attribute in the reference. The function executes
transformations specified by the \code{tx} field (before calculation of \code{y}) and
the \code{yt} field (after calculating of \code{y}), so that the \code{y} have the original
units even if the reference is based on some transform of \code{y}.

Functions \code{y2z()} and \code{z2y()} functions are the inverse of each other.
}
\examples{
# height centiles SD -2:2 according to WHO standard for 6 month year old girl
# using built-in reference
z2y(z = -2:2, x = rep(0.5, 5), "who_2006_hgt_female_", pkg = "centile")

# same using external Dutch reference
fn <- system.file("testdata/nl_2009_hgt_female_nl.txt", package = "centile")
myref <- import_rif(fn)
head(myref)
head(attr(myref, "study"))
z2y(z = -2:2, x = rep(0.5, 5), myref)

# find location of -2 SD line for 6 months old boys and girls, WHO Standard
z2y(c(-2, -2), c(0.5, 0.5), c("who_2006_hgt_male_", "who_2006_hgt_female_"))
}
\author{
Stef van Buuren, 2021
}
